Migrate navigation from app to shared package
