+++ design-system/src/commonMain/kotlin/com/lurian/design_system/components/theme/Color.kt
@@ -7,3 +7,11 @@ val DarkGray = Color(0xFF121212)
 val LightGray = Color(0xFFBDBDBD)
 val White = Color(0xFFFFFFFF)
 val Black = Color(0xFF000000)
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
